Use a fixed time to reset
Because, this has an indirect relationship with the low light automation. It needs to reset at the same time the conditions of that automation meet, which is at 08:00, otherwise on a morning one wants to skip the wake-up lights, as soon as this switch resets, the low light automation might run.