Bugfix: Unintended consequence of the hysteresis/dead bands
The issue is that if there's a large change, say from 4 kWh (pessimistic) to 20 kWh (very optimistic), then it will default to pessimistic. The original idea was to be robust against jitter of the forecast (mostly due to Forecast.solar being pretty bad). My average sensor and use of Open Meteo make things much better. Let's see how it goes.